+++ HISTORY.rst
@@ -3,7 +3,16 @@
 History
 -------
 
-0.1.0 (2014-07-02)
----------------------
+0.2.0 (2014-07-04)
+==================
 
-* First release on PyPI.
+* Add tests for `box` and `monster`
+* Create `box` module to contain server code
+* Add Sphinx-compatible docstrings to all code
+* Create framework for `box` to parse commands
+* Add `export_monster` method to `Monster` to make export easier
+
+0.1.0 (2014-07-02)
+==================
+
+* Initial release.